changed

+++ src/main.rs
@@ -12,7 +12,7 @@ fn main() {
     let repeat = 2 as i32;
     let root_process = world.process_at_rank(root_rank);
 
-    let mut x = vec![0;repeat];
+    let mut x = vec![0;repeat as usize];
     if rank == root_rank {
         let v = (0..size*repeat).collect::<Vec<_>>();
         root_process.scatter_into_root(&v[(rank*repeat) as usize..((rank+1)*repeat) as usize], &mut x[..]);
